Move events and actions types to separated files for maintainability
